requirements.txt
Django==1.7.1
Jinja2==2.7.3
certifi==14.05.14
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
python-dateutil==2.2
mock==1.0.1
psycopg2==2.5.4
requests==2.4.0
selenium==2.44.0
static3==0.5.1
